refactor: Moved everything possible into $lib
