+++ src/lib/components/buttons/show-directory.svelte
