+++ src/lib/components/cards/pattern-settings.svelte
