+++ src/lib/components/search-box/input-type.svelte
