+++ src/lib/components/tables/actions/pattern-actions.svelte
