+++ src/lib/utils/fabric.ts
