+++ src/routes/contexts/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import Table from "./table/context-table.svelte";
-  import TextBox from "./edit/text-box.svelte";
+  import Table from "../../lib/components/tables/contexts.svelte";
+  import TextBox from "../../lib/components/text-area/edit-content.svelte";
   import * as Resizable from "$lib/components/ui/resizable";
   import { writable } from "svelte/store";
