+++ src/routes/settings/+page.svelte
@@ -10,8 +10,8 @@
 
   // tables
   // import ModelsTable from "$lib/components/ModelsTable.svelte";
-  import SecretsTable from "./ApiKeysTable.svelte";
-  import BaseUrlTable from "./BaseUrlTable.svelte";
+  import SecretsTable from "../../lib/components/tables/api-keys.svelte";
+  import BaseUrlTable from "../../lib/components/tables/base-urls.svelte";
 
   // tauri plugins
   import { invoke } from "@tauri-apps/api/core";
